Migrate physical plan tests
